ARM64: dts: MSM8996: import Energy Model
This EM is from https://github.com/EAS-Project, which is more
conservative than Google's marlin EM. We can use schedtune to boost
as necessary to increase performance but this will provide some
nice battery savings.

As for yoshino, we plan to further optimise by testing power output
per freq on-device.

Signed-off-by: Laster K. (lazerl0rd) <officiallazerl0rd@gmail.com>