Makefile: ignore Polly optimisations when built in-line with AOSP
Having these on causes a load of issues when building AOSP in-line, so ignore the
config in those circumstances. I selected PLATFORM_VERSION_CODENAME as the variable
to look out for as TARGET_BUILD_VARIANT can be sometimes defined in QCALCD, thanks
to Danny Lin on picking up that one.

Suggested-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: Laster K. (lazerl0rd) <officiallazerl0rd@gmail.com>